server/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==3.2.23
django-allauth==0.58.2
django-enumfields==2.1.1
django-jsonfield==1.4.1
djangorestframework==3.14.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
